--- drush/Commands/PolicyCommands.php
@@ -1,42 +1,0 @@
-<?php
-
-namespace Drush\Commands;
-
-use Consolidation\AnnotatedCommand\CommandData;
-
-/**
- * General policy commands and hooks for the application.
- */
-class PolicyCommands extends DrushCommands {
-
-  /**
-   * Overwrite the sites.local.php file prior to starting the server.
-   *
-   * @hook pre-command runserver
-   */
-  public function preRunserver(CommandData $commandData) {
-    $uri = $commandData->input()->getOption('uri');
-
-    $data = <<<EOD
-<?php
-/**
- * @file
- * Automatically generated by PolicyCommands.
- */
- 
-\$sites['8888.localhost'] = '$uri';
-EOD;
-
-    if ($uri) {
-      file_put_contents(DRUPAL_ROOT . '/sites/sites.local.php', $data);
-      $this->logger()->notice(dt('Overwrote sites.local.php to route @uri site.', [
-        '@uri' => $uri,
-      ]));
-    }
-    else {
-      file_put_contents(DRUPAL_ROOT . '/sites/sites.local.php', "<?php\n");
-      $this->logger()->notice(dt('Overwrote sites.local.php to route default site.'));
-    }
-  }
-
-}
